% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_connect_sgbd.R
\name{connect_sgbd}
\alias{connect_sgbd}
\title{Função que cria a conexão com o Banco de Dados}
\usage{
connect_sgbd(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Função que cria a conexão com o Banco de Dados
}
