% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_connect_sgbd.R
\name{criar_bd}
\alias{criar_bd}
\title{Função que cria o Banco de Dados}
\usage{
criar_bd(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Função que cria o Banco de Dados
}
