% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_criar_diretorios.R
\name{criar_diretorios}
\alias{criar_diretorios}
\title{Função que cria os diretório do Web Scraping}
\usage{
criar_diretorios()
}
\description{
Os diretórios do Web Scraping são criados no diretório de trabalho escolhido pelo usuário
}
