% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_criar_tb_dcalendario.R
\name{criar_tb_dcalendario}
\alias{criar_tb_dcalendario}
\title{Função que gera a tabela 'dCalendario' com a relação de meses e ano para o Web Scraping;}
\usage{
criar_tb_dcalendario(ano_inicio, sgbd = "sqlite")
}
\arguments{
\item{ano_inicio}{Define o ano de início do Web Scraping}

\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Função que gera a tabela 'dCalendario' com a relação de meses e ano para o Web Scraping;
}
