% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_criar_tb_dmunicipios.R
\name{criar_tb_dmunicipios}
\alias{criar_tb_dmunicipios}
\title{Função que gera a tabela 'tabela_tcm_dmunicipios' no Banco de Dados}
\usage{
criar_tb_dmunicipios(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado.
Por padrão, é definido como sqlite}
}
\description{
Função que gera a tabela 'tabela_tcm_dmunicipios' no Banco de Dados
}
