% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_criar_tb_dmunicipios_entidades.R
\name{criar_tb_dmunicipios_entidades}
\alias{criar_tb_dmunicipios_entidades}
\title{Função que gera a tabela "tabela_tcm_dmunicipios_entidades" no Banco de Dados}
\usage{
criar_tb_dmunicipios_entidades(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado.
Por padrão, é definido como sqlite}
}
\description{
Função que gera a tabela "tabela_tcm_dmunicipios_entidades" no Banco de Dados
}
