% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_criar_tb_entidades_alvos_paginas.R
\name{criar_tb_entidades_alvos_paginas}
\alias{criar_tb_entidades_alvos_paginas}
\title{Função que gera os dados das entidades alvos que servirão de parâmetro
para as requisições das páginas}
\usage{
criar_tb_entidades_alvos_paginas(cod_municipios_alvos, sgbd)
}
\arguments{
\item{cod_municipios_alvos}{Código dos Municípios alvos}

\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}
}
\description{
Função que gera os dados das entidades alvos que servirão de parâmetro
para as requisições das páginas
}
