% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_criar_tb_requisicoes_despesas.R
\name{criar_tb_requisicoes_despesas}
\alias{criar_tb_requisicoes_despesas}
\title{Função que gera a Tabela de Requisições "tabela_paginas_links"}
\usage{
criar_tb_requisicoes_despesas(sgbd)
}
\arguments{
\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}
}
\description{
Função que gera a Tabela de Requisições "tabela_paginas_links"
}
