% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_backup_arquivos.R
\name{executar_backup_arquivos}
\alias{executar_backup_arquivos}
\title{Função que realiza o backup dos arquivos}
\usage{
executar_backup_arquivos(backup_local = "SIM", backup_nuvem = "NAO")
}
\arguments{
\item{backup_local}{É definido como "SIM" como padrão para realizar
o Backup do Banco de Dados e dos arquivos HTML e CSV.
Mas pode ser marcado como "NAO".}

\item{backup_nuvem}{É definido como "NAO" como padrão.
Mas pode ser marcado como "SIM", caso deseje realizar
o backup no DroBox.}
}
\description{
Função que executa o backup do Banco de Dados, dos arquivos HTML,
e dos arquivos CSVs no ambiente local e na nuvem (DropBox).
Obs: É preciso configurar o token do DropBox para a função funcionar.
Saiba como gerar o seu Token em: \href{https://github.com/karthik/rdrop2}{linked phrase}
}
