% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_executar_data_wrangling_html_despesas.R
\name{executar_data_wrangling_html_despesas}
\alias{executar_data_wrangling_html_despesas}
\title{Função que realiza a extração de dados das despesas do arquivo html}
\usage{
executar_data_wrangling_html_despesas(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}
}
\description{
Função que realiza a extração de dados de despesas que estao no arquivo html
}
