% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_executar_scraping_num_pags.R
\name{executar_scraping_num_pags}
\alias{executar_scraping_num_pags}
\title{Função que executa o Web Scraping das Páginas do TCM-ba}
\usage{
executar_scraping_num_pags(sgbd)
}
\arguments{
\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}
}
\description{
Função que executa o Web Scraping das Páginas do TCM-ba
}
