% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_executar_tidy_data.R
\name{executar_tidy_data}
\alias{executar_tidy_data}
\title{Função que transforma os dados para o padrão "tidy data"}
\usage{
executar_tidy_data(sgbd)
}
\arguments{
\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}
}
\description{
Função que transforma os dados para o padrão "tidy data"
}
