% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_executar_web_scraping.R
\name{executar_web_scraping}
\alias{executar_web_scraping}
\title{Função que executa o Web Scraping}
\usage{
executar_web_scraping(nome_scraping = "ws_tcmba_despesas",
  ano_inicio = "2019", cod_municipios_alvos = c(2928703),
  sgbd = "sqlite", repetir = "SIM", backup_local = "SIM",
  backup_nuvem = "NAO", exportar_nuvem = "NAO")
}
\arguments{
\item{nome_scraping}{Nome do Diretório que será criado para alocar os dados do Web Scraping}

\item{ano_inicio}{Exercíco (ano) inicial da coleta de dados}

\item{cod_municipios_alvos}{Código do Município alvo}

\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}

\item{repetir}{É definido "SIM" como padrão. Mas pode ser marcado como "NAO",
caso não deseje repetir (na próxima execução) as consulta do Web Scraping
que falharam ou que não foram identificadas resposta do ente municipal
no dia de execução do Web Scraping}

\item{backup_local}{É definido como "SIM" como padrão para realizar
o Backup do Banco de Dados e dos arquivos HTML e CSV.
Mas pode ser marcado como "NAO".}

\item{backup_nuvem}{Realiza o Backup dos dados para o DropBox. Como padrão, é definido "NAO".
Obs: É preciso configurar o token do DropBox antes de executar
(ver instruções na função 'exportar_csv_dropbox').}

\item{exportar_nuvem}{É definido como "NAO" como padrão para realizar
Exportar os CSVs para o DropBox com o objetivo de conectar o Power BI}
}
\description{
Essa função foi desenvolvida para executar todas as etapas do Web Scraping
e, então, coletar os dados públicos das Despesas dos Municípios do Estado da Bahia
custodiados no site do TCM-Ba
}
\examples{
\dontrun{
executar_web_scraping(nome_scraping = "ws_tcmba_despesas",
                      ano_inicio = 2019,
                      cod_municipios_alvos = c(2928703),
                      sgbd = "sqlite",
                      repetir = "SIM",
                      backup_local = "SIM",
                      backup_nuvem = "NAO",
                      exportar_nuvem = "NAO")
}


}
