% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_exportar_dropbox.R
\name{exportar_csv_dropbox}
\alias{exportar_csv_dropbox}
\title{Função que exporta os arquivos CSV para o DropBox, com o objetivo
conectar com o Power BI}
\usage{
exportar_csv_dropbox(exportar_nuvem = "NAO")
}
\arguments{
\item{exportar_nuvem}{É definido como "NAO" como padrão. Mas pode ser marcado como "SIM"
para realizar a Exportação do arquivo .zip (com diversos CSVs) para o DropBox.}
}
\description{
Obs: É preciso configurar o token do DropBox para a função funcionar.
Saiba como gerar o seu Token em: \href{https://github.com/karthik/rdrop2}{linked phrase}
}
