% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_utils.R
\name{gravar_erro}
\alias{gravar_erro}
\title{Função para gravar erro da execução do código}
\usage{
gravar_erro(log_request, nm_log_erro = "", entrada = "", id = "",
  cod_entidade = "", nm_entidade = "", ano = "", mes = "",
  outros = "", sgbd = "sqlite")
}
\arguments{
\item{log_request}{Valor gerado pela função log_data_hora()}

\item{nm_log_erro}{Nome do erro atribuído pelo desenvolvedor;}

\item{entrada}{Resultado após execução da função envolvida pela função purrr::safely}

\item{id}{Número do ID no Banco de Dados;}

\item{cod_entidade}{Código da Entidade Municipal do Web Scraping;}

\item{nm_entidade}{Nome da Entidade Municipal do Web Scraping;}

\item{ano}{Ano de referência da Entidade Municipal;}

\item{mes}{Mês de referência da Entidade Municipal;}

\item{outros}{Outras informações adicionadas pelo Desenvolvedor;}

\item{sgbd}{Nome do Banco de Dados usado na execução do Código}
}
\description{
Função para gravar erro da execução do código
}
