% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_utils.R
\name{limpar_nome}
\alias{limpar_nome}
\title{Função para limpar nomes de caracteres especiais}
\usage{
limpar_nome(x)
}
\arguments{
\item{x}{string}
}
\description{
Limpa o nome dos entes municipais para retirar caracteres especiais
}
