% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_utils.R
\name{log_data_hora}
\alias{log_data_hora}
\title{Função para registrar o fuso horário}
\usage{
log_data_hora()
}
\description{
Função usada para registrar o fuso horário do Brasil
mesmo que o servidor tenha o fuso horário de outro país
}
