% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcm_cod_municipios}
\alias{tcm_cod_municipios}
\title{Codigo dos Municipios}
\format{A data frame with 996 rows and 4 variables:
\describe{
\item{cod_municipio}{Codigo do Municipio}
\item{nm_municipio}{Nome do Municipio}
\item{cod_entidade}{Codigo do ente municipal - Nao usar esse numero do Web Scraping}
\item{nm_entidade}{Codigo do ente municipal}
...
}}
\source{
\url{http://www.tcm.ba.gov.br/}
}
\usage{
tcm_cod_municipios
}
\description{
Dataset contendo o código dos municípios para servir de referência no Web Scrapingstr()
}
\keyword{datasets}
