% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_utils.R
\name{valor_monetario}
\alias{valor_monetario}
\title{Função para padronizar valores monetários}
\usage{
valor_monetario(x)
}
\arguments{
\item{x}{string}
}
\description{
Limpa os valores monetários de caracteres especiais.
Ex: 'R$ 1.059,00' para '1059.00'
}
