% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_executar_scraping_html_despesas.R
\name{executar_scraping_html_despesas}
\alias{executar_scraping_html_despesas}
\title{Função que executa o Web Scraping das Páginas de despesas do TCM-ba}
\usage{
executar_scraping_html_despesas(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite.}
}
\description{
Função que executa o Web Scraping das Páginas de despesas do TCM-ba
}
